Update to json-refs 2.1.
The old json-refs is missing a lot of functionality that the new one
provides.  It actually makes it a lot easier to use as well.  I took the
code being used here from the CLI command that json-refs now ships with
so things will work with gulp in much the same way as they work with the
json-refs CLI.